+++ server/initializers/database.ts
@@ -37,6 +37,8 @@ import { VideoStreamingPlaylistModel } from '../models/video/video-streaming-pla
 import { VideoPlaylistModel } from '../models/video/video-playlist'
 import { VideoPlaylistElementModel } from '../models/video/video-playlist-element'
 import { ThumbnailModel } from '../models/video/thumbnail'
+import { PremiumStoragePlanModel } from '../models/premium-storage-plan'
+import { userPremiumStoragePaymentModel } from '../models/user-premium-storage-payments'
 import { PluginModel } from '../models/server/plugin'
 import { QueryTypes, Transaction } from 'sequelize'
 
@@ -109,6 +111,8 @@ async function initDatabaseModels (silent: boolean) {
     VideoPlaylistModel,
     VideoPlaylistElementModel,
     ThumbnailModel,
+    PremiumStoragePlanModel,
+    userPremiumStoragePaymentModel,
     PluginModel
   ])