+++ shared/models/server/premium-storage-plan-interface.ts
@@ -1,0 +1,9 @@
+export interface interfacePremiumStoragePlan {
+  id?: number
+  name: string
+  quota: number
+  dailyQuota: number
+  priceTube: number
+  duration: number
+  active: boolean
+}