+++ shared/models/server/server-config.model.ts
@@ -161,4 +161,8 @@ export interface ServerConfig {
       }
     }
   }
+
+  premium_storage: {
+    enabled: boolean
+  }
 }